third commit

+++ src/main.c
@@ -49,7 +49,7 @@ bool is_border_correct(char *border){
     if (border == NULL) {
         return false;
     }
-    for(int i = 0; i < strlen(border); i++){
+    for(size_t i = 0; i < strlen(border); i++){
         if(!is_a_number( border[i])){
             return false;
         }
